Don't try to access a long field as an int field.
I've raised a bug to enhance CheckJNI. For now, this was caught by a
setSoTimeout test.

Bug: 5159133

+++ luni/src/main/native/libcore_io_Posix.cpp
@@ -1116,8 +1116,8 @@ static void Posix_setsockoptTimeval(JNIEnv* env, jobject, jobject javaFd, jint l
     static jfieldID tvUsecFid = env->GetFieldID(JniConstants::structTimevalClass, "tv_usec", "J");
     int fd = jniGetFDFromFileDescriptor(env, javaFd);
     struct timeval value;
-    value.tv_sec = env->GetIntField(javaTimeval, tvSecFid);
-    value.tv_usec = env->GetIntField(javaTimeval, tvUsecFid);
+    value.tv_sec = env->GetLongField(javaTimeval, tvSecFid);
+    value.tv_usec = env->GetLongField(javaTimeval, tvUsecFid);
     throwIfMinusOne(env, "setsockopt", TEMP_FAILURE_RETRY(setsockopt(fd, level, option, &value, sizeof(value))));
 }
